% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_weak.R
\name{accuracy_coverage_density}
\alias{accuracy_coverage_density}
\title{Probability density of true label and observed weak labels given accuracy and coverage}
\usage{
accuracy_coverage_density(lambda, y, a = rep(0.8, length(lambda)),
  b = rep(0.8, length(lambda)), grad = FALSE)
}
\arguments{
\item{lambda}{An mx1 vector of weak labels in \code{{-1,0,1}}.}

\item{y}{A scalar ground truth label in \code{{-1,1}} Cannot be missing.}

\item{a}{An mx1 vector of accuracy estimates in \code{[0,1]} for each labelling function.}

\item{b}{An mx1 vector of coverage estimates in \code{[0,1]} for each labelling function.}

\item{grad}{A boolean flag. If TRUE, the gradient of the density is computed instead.}
}
\value{
The probability density of an observed set of weak labels, given parameter values for accuracy and coverage.
}
\description{
This function represents the basic generative model of weak supervision, where the observed values of a set of labelling functions are related via a simple binomial model to estimates of accuracy and coverage for each labelling function.
}

