% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_classifier.R
\name{fit_logistic}
\alias{fit_logistic}
\title{A simple wrapper for the glm function}
\usage{
fit_logistic(data, weights = NULL)
}
\arguments{
\item{data}{An nx(p+1) matrix, with one column called "Y" containing the labels, and p columns containing the covariates (categorical or numeric).}

\item{weights}{(optional) An nx1 vector of weights in \code{[0,1]} representing the accuracy of each label (formally the number of trials in that binomial experiment).}
}
\value{
A fitted glm object.
}
\description{
This wrapper simplifies the interface of the glm function to that of a logistic classifier. It also abuses the inteface of the glm a little bit, using the optional parameter 'weights' which formally corresponds to the number of trials in a binomial experiment as a proxy for an estimated accuracy of that particular label. This is a theoretically sensible move, which allows us to use existing interfaces for logistic regression in many languages, but can be improved with a re-implemented gradient descent.
}

