% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_simulate.R
\name{is_valid_label}
\alias{is_valid_label}
\title{Helper function to check label space}
\usage{
is_valid_label(y, label_space = c(-1, 0, 1))
}
\arguments{
\item{y}{A vector, matrix, or data frame of labels}

\item{label_space}{A vector of allowed labels}
}
\value{
TRUE/FALSE
}
\description{
This function checks if all the values in a data frame, matrix or vector belong in the given space of labels. It is used for testing.
}
\examples{
stopifnot(!is_valid_label(c(0,1,-1), c(-1,1)))
stopifnot(is_valid_label(c(0,1,-1), c(-1,0,1)))
}

