% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_weak.R
\name{marginal_log_likelihood}
\alias{marginal_log_likelihood}
\title{Marginal log likelihood gradient}
\usage{
marginal_log_likelihood(Lambda, a = rep(0.8, dim(Lambda)[2]), b = rep(0.8,
  dim(Lambda)[2]))
}
\arguments{
\item{Lambda}{An nxm matrix of weak labels, where each row respresents the weak labels produced by m labelling functions on a single data example.}

\item{a}{An mx1 estimate of accuracy probabilities.}

\item{b}{An mx1 estimate of coverage probabilities}
}
\value{
A scalar value equal to the marginal log likelihood.
}
\description{
This function computes the marginal log likelihood of a set of weak labels given an estimate of accuracy and coverage for each labelling function.
}

