% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_weak.R
\name{marginal_log_likelihood_grad}
\alias{marginal_log_likelihood_grad}
\title{Marginal log likelihood gradient}
\usage{
marginal_log_likelihood_grad(Lambda, a = rep(0.8, dim(Lambda)[2]),
  b = rep(0.8, dim(Lambda)[2]))
}
\arguments{
\item{Lambda}{An nxm matrix of weak labels, where each row respresents the weak labels produced by m labelling functions on a single data example.}

\item{a}{An mx1 estimate of accuracy probabilities.}

\item{b}{An mx1 estimate of coverage probabilities}
}
\value{
An mx1 vector where the i'th element represents the partial derivative of the marginal log likelihood with respect to the accuracy of the i'th labelling function, evaluated at the currently provided values a.
}
\description{
This function computes the gradient of the marginal log likelihood of a set of weak labels with respect to the accuracy vector (the coverage vector can be estimated in closed-form).
}

