% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_weak.R
\name{plot_accuracy_estimates}
\alias{plot_accuracy_estimates}
\title{Plotting routine for accuracy estimates}
\usage{
plot_accuracy_estimates(traj, a_true, main = "Accuracy Estimates")
}
\arguments{
\item{traj}{A matrix containing the trajectories of the estimates of accuracy during gradient descent}

\item{a_true}{The true accuracy values}

\item{main}{A title for the plot}
}
\description{
A convenience function for plotting accuracy trajectories
}

