% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_classifier.R
\name{predict_logistic}
\alias{predict_logistic}
\title{A simple wrapper for the predict method of a glm object}
\usage{
predict_logistic(m_obj, newdata)
}
\arguments{
\item{m_obj}{A fitted glm object}

\item{newdata}{A nxp matrix with columns identically named to the covariates that were used to train the model. The matrix can have additional columns without having an effect on the output.}
}
\value{
A list with two elements: \code{scores} is an nx1 vector of probabilities of the respective label being equal to 1, and \code{y_hat} are the estimated labels with 0.5 as the decision boundary.
}
\description{
This wrapper simplifies the interface of the predict method of a glm object to that of a logistic classifier.
}

