% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_weak.R
\name{run_experiment}
\alias{run_experiment}
\title{Run an entire marginal log likelihood optimisation experiment}
\usage{
run_experiment(accuracy, coverage, ground_truth = FALSE, n = 1000, p = 10,
  maxiters = 200, stepsize = 0.5)
}
\arguments{
\item{accuracy}{the desired accuracy for each labelling rule}

\item{coverage}{the desired coverage for each labelling rule}

\item{ground_truth}{whether to allow the optimiser to see ground truth or not}

\item{n}{The number of examples to generate}

\item{p}{The number of covariates to generate}

\item{maxiters}{The maximum number of iterations for the gradient descent loop}

\item{stepsize}{The step size for the gradient descent loop}
}
\description{
A convenience function to run experiments with different accuracy and coverage parameters and assess the performance of the optimisation routine
}

