% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_simulate.R
\name{simulate_logreg}
\alias{simulate_logreg}
\title{Simulate data from a logistic regression}
\usage{
simulate_logreg(ntrain = 1000, ntest = ntrain, p = 2, betav = 5,
  seed = 2)
}
\arguments{
\item{ntrain}{The number of datapoints in the train dataset}

\item{ntest}{The number of datapoints in the test dataset}

\item{p}{The number of covariates (excludin the intercept)}

\item{betav}{The variance of the prior for the regression coefficients}

\item{seed}{Always set your seed :)}
}
\value{
A list of two data frames. The first data frame is called \code{data} and has dimension \code{ntrain x (p+1)}. The second data frame is called \code{newdata} and has dimension \code{ntest x (p+1)}. In both cases, the label space is \code{{0,1}} (as per standard convention for logistic regression), and the labels are in the last column of the respective data frames, called \code{Y}.
}
\description{
This function simulates two datasets from a logistic regression of dimension p: one training dataset, and a test dataset. The coefficients are normally distributed as \code{N(0,betav)} where \code{betav} can be specified by the user. An intercept is added.
}
\examples{
simulate_logreg(ntrain=100)
}

