% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library_weak.R
\name{transform_weak_data}
\alias{transform_weak_data}
\title{A transformation from weakly labelled examples to weighted examples}
\usage{
transform_weak_data(X, Lambda, accuracies)
}
\arguments{
\item{X}{is an nxp matrix of data examples - note this should not include the ground truth label}

\item{Lambda}{is the matrix of weak labels}

\item{accuracies}{is a vector of estimated (or known) accuracies for each labelling rule. This can include perfect accuracy in the presence of ground truth.}
}
\description{
It is possible to map weak labels to weighted learning, and in this fashion perform noise-aware empirical loss minimisation while using existing classifier interfaces
}

